requirements.txt
fastapi==0.95.1
pydantic==1.10.7
uvicorn==0.22.0
openai==0.27.0
torch==2.0.1
nltk==3.8.1
spacy==3.5.0
pyspellchecker==0.7.2
transformers==4.30.2
joblib==1.2.0
pandas==2.0.3